Euler quaternion

+++ cubittest/cubit/math/Matrix4Test.cpp
@@ -30,7 +30,7 @@ TEST(Matrix4Test, elementsConstructor_equalsElement) {
   int k = 1;
   for (int i = 0; i < 4; i++) {
     for (int j = 0; j < 4; j++) {
-      EXPECT_FLOAT_EQ(k, m[i][j]);
+      EXPECT_FLOAT_EQ((float)k, m[i][j]);
       k++;
     }
   }
